Extract map_figure callbacks
